FEAT : redis 패키지 재배치 및 재적용
