+++ concert-reservation-api/src/main/java/io/hhplus/concert/application/user/scheduler/UserQueueScheduler.java
@@ -1,7 +1,6 @@
 package io.hhplus.concert.application.user.scheduler;
 
-import io.hhplus.concert.application.user.UserQueueService;
-import io.hhplus.concert.infrastructure.repository.user.RedisQueuePublisher;
+import io.hhplus.concert.application.messaging.CheckAndPublishMessageUseCase;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -12,20 +11,16 @@ public class UserQueueScheduler {
 
     private static final Logger logger = LoggerFactory.getLogger(UserQueueScheduler.class);
 
-    private final RedisQueuePublisher redisQueuePublisher;
+    private final CheckAndPublishMessageUseCase checkAndPublishMessageUseCase;
 
-    public UserQueueScheduler(RedisQueuePublisher redisQueuePublisher) {
-        this.redisQueuePublisher = redisQueuePublisher;
+
+    public UserQueueScheduler(CheckAndPublishMessageUseCase checkAndPublishMessageUseCase) {
+        this.checkAndPublishMessageUseCase = checkAndPublishMessageUseCase;
     }
 
     @Scheduled(fixedRate = 60000)
     public void execute() {
-        try {
-            redisQueuePublisher.publish("Test Message from UserService");
-            logger.info("UserQueueScheduler completed execution successfully");
-        } catch (Exception e) {
-            logger.error("Error during UserQueueScheduler execution", e);
-        }
+        checkAndPublishMessageUseCase.execute("Activate next user");
     }
 
 }
