+++ concert-reservation-api/src/main/java/io/hhplus/concert/domain/messaging/redis/RedisPublisher.java
@@ -1,0 +1,7 @@
+package io.hhplus.concert.domain.messaging.redis;
+
+public interface RedisPublisher {
+
+    void publish(String message);
+
+}
