+++ concert-reservation-api/src/main/java/io/hhplus/concert/interfaces/messaging/redis/RedisSubscriber.java
@@ -1,5 +1,6 @@
-package io.hhplus.concert.application.user;
+package io.hhplus.concert.interfaces.messaging.redis;
 
+import io.hhplus.concert.application.user.UserQueueService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.redis.connection.Message;
@@ -7,13 +8,13 @@ import org.springframework.data.redis.connection.MessageListener;
 import org.springframework.stereotype.Component;
 
 @Component
-public class RedisQueueSubscriber implements MessageListener {
+public class RedisSubscriber implements MessageListener {
 
-    private static final Logger logger = LoggerFactory.getLogger(RedisQueueSubscriber.class);
+    private static final Logger logger = LoggerFactory.getLogger(RedisSubscriber.class);
 
     private final UserQueueService userQueueService;
 
-    public RedisQueueSubscriber(UserQueueService userQueueService) {
+    public RedisSubscriber(UserQueueService userQueueService) {
         this.userQueueService = userQueueService;
     }
